% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logiBin.R
\docType{data}
\name{loanData}
\alias{loanData}
\title{Simulated default data of 100 customers}
\format{A data frame with 100 rows and 6 variables:
\describe{
  \item{bad_flag}{Indicates whether a customer has defaulted or not}
  \item{age}{Age of the customer}
  \item{LTV}{Ratio of amount of loan to amount of collateral}
  \item{location}{Indicates the location of the customer}
  \item{balance}{Account balance}
  \item{score}{credit score of the customer}
}}
\usage{
loanData
}
\description{
A dataset containing simulated data about the characteristic of a customer applying for a loan. The dependent variable is "bad_flag" which indicates whether the customer defaults or not.
}
\keyword{datasets}
