% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logiBin.R
\name{manualSplit}
\alias{manualSplit}
\title{Split a variable based on specified cuts}
\usage{
manualSplit(binObj, splitVar, y, splits, df)
}
\arguments{
\item{binObj}{- An object returned by getBins or any other function (except createBins) in this package}

\item{splitVar}{- The name of the variable that has to be split}

\item{y}{- The dependent variable}

\item{splits}{- The splits for the variable}

\item{df}{- A data frame}
}
\value{
Returns a list containing 3 objects. Similar to the getBins function
}
\description{
This function splits variables based on cuts that have been input manually
}
\examples{
b1 <- getBins(loanData, "bad_flag", c('age', 'score', 'balance'), minCr=0.01)
b1 <- manualSplit(b1, 'age', 'bad_flag', c(25,40,55), loanData)

}
